Selection works in LazyTreeView.
MenuView no longer uses undefined/null as a proper selected value. This was to make a case where a SelectButtonView originally has undefined as the selected value not to show all submenus (as they have undefined as their value) as selected.
I feel this may break something somewhere where we've used null as a proper value. A better way to fix this would be to give the menu the selected value in constructor and have the menu's selection wrapped in a object that explicitly states whether or not there is a selection and what is its value.
As that would require changes everywhere the menu is used, I did not do it at this point.